package restructure, code tidy up
