+++ cmd/ccexplorer/ccexplorer.go
@@ -1,17 +1,21 @@
 package main
 
 import (
-	"github.com/cduggn/ccexplorer/internal/core/handlers/commandline"
-	"github.com/cduggn/ccexplorer/internal/core/logger"
+	"github.com/cduggn/ccexplorer/internal/cli-new"
+	"github.com/sagikazarmark/slog-shim"
 	"os"
 )
 
 func main() {
-	root := commandline.RootCommand()
+	root := cli_new.RootCommand()
 
 	if err := root.Execute(); err != nil {
-		logger.ErrorOut(err)
+		slog.Error("error", ErrAttr(err))
 		os.Exit(126)
 	}
 
+}
+
+func ErrAttr(err error) slog.Attr {
+	return slog.Any("error", err)
 }
