+++ internal/aws-new/config.go
@@ -1,0 +1,12 @@
+package aws_new
+
+import "github.com/spf13/viper"
+
+func Profile() string {
+	awsProfile := viper.GetString("aws_profile")
+	if awsProfile == "" {
+		awsProfile = "not-provided"
+	}
+
+	return awsProfile
+}
